Rename chapter to supplementary to match design and for consistency; rename intro to be supplementaryStatementsIntro to avoid confusion
